neural_worker/requirements.txt
celery==5.3.4
Pillow
redis
ultralytics
yt-dlp
efficientnet_pytorch